+++ src/RampGenerator.h
@@ -3,7 +3,7 @@
 
 #include "FastAccelStepper.h"
 #include "RampCalculator.h"
-#include "RampConstAcceleration.h"
+#include "RampControl.h"
 #include "fas_arch/common.h"
 
 class FastAccelStepper;